AUTHING IS DONE
I finnally have the auth working in C# but that also means I had to
change my database schema a bit in order to allow for the way auth
actually works and not how I thought or how the web api led me to
believe it did but no matter it is working and all is well for now.

+++ Database/Tables/dbo.Spotify_Playlist_Track_Changes.Table.sql
@@ -8,7 +8,7 @@ BEGIN
   --Track
 	,TrackID BIGINT NOT NULL
   --Position old and new :)
-    ,Position BIGINT NOT NULL
+    ,Position INT NOT NULL
     ,ChangeTypeID INT NOT NULL
     ,FOREIGN KEY(TrackID) REFERENCES dbo.Spotify_Track(TrackID)
 	,FOREIGN KEY(PlaylistID) REFERENCES dbo.Spotify_Playlist(PlaylistID)
